+++ e2e/rstudio/utils/ace.ts
@@ -76,6 +76,7 @@ export namespace Ace {
     find(needle: string): void;
     insert(text: string): void;
     navigateLineEnd(): void;
+    navigateFileEnd(): void;
     selectAll(): void;
     execCommand(name: string): void;
     exitMultiSelectMode(): void;